javadoc, cleanup

+++ library/src/wei/mark/standout/Utils.java
@@ -1,7 +1,7 @@
 package wei.mark.standout;
 
 public class Utils {
-	public static boolean isSet(long flags, long flag) {
+	public static boolean isSet(int flags, int flag) {
 		return (flags & flag) == flag;
 	}
 }
